+++ ep-sdk/src/tasks/EpSdkEventApiVersionTask.ts
@@ -51,7 +51,7 @@ export class EpSdkEventApiVersionTask extends EpSdkVersionTask {
     epObjectExists: false  
   };
   private readonly Default_TEpSdkEventApiVersionTask_Settings: Partial<TEpSdkEventApiVersionTask_Settings> = {
-    description: `Created by ${EpSdkConfig.getAppName()}.`,
+    // description: `Created by ${EpSdkConfig.getAppName()}.`,
   }
   private getTaskConfig(): IEpSdkEventApiVersionTask_Config { 
     return this.epSdkTask_Config as IEpSdkEventApiVersionTask_Config;